Move helper functions & decorators to cli_util.py module
